+++ CHANGELOG.md
@@ -12,7 +12,8 @@
 
 * Refactor `hexapdf modify` command into three individual commands `modify`,
   `merge` and `optimize`
-* Rename `hexapdf extract` to `hexapdf files`
+* Rename `hexapdf extract` to `hexapdf files` and the option `--indices` to
+  `--extract`
 * Show PDF trailer by default with `hexapdf inspect`
 * Refactor CLI command classes to use specialized superclass
   [HexaPDF::CLI::Command]